+++ gameServices/ygo/YGOCardService.py
@@ -10,7 +10,7 @@ IMAGE_EXTENSION = ".jpg"
 def getCardsFromFile(totalCards):
     cardImages = []
     cardCount = 0
-    with open(Utils.getPath('input', 'ygoInput.txt'), 'r') as f:
+    with open(Utils.getPath('../../input', 'ygoInput.txt'), 'r') as f:
         lines = f.readlines()
         f.close()
     for line in lines:
